Replaced per-GO sprite rendering with a wrapper
Wrapper takes in a image source (string) and a transform (ptr), reacts to rotation/pos changes. The origin of the sprite is set to the centre point of the supplied image source's texture rect

TODO: Allow GOs to have parents. so I can add wheels as their own independent items